Added S3 service, fixed some essues in user services, added Vehicle handling (ability to publish and delete vehicles), added minor changes in Firebase service controllers and silent pushing

+++ src/main/java/world/horosho/CarMeeter/DB/Repositories/FCMRepository.java
@@ -5,5 +5,5 @@ import reactor.core.publisher.Flux;
 import world.horosho.CarMeeter.DB.Models.POST.FCM;
 
 public interface FCMRepository extends ReactiveCrudRepository<FCM, Long> {
-    Flux<FCM> findByUserId(int id);
+    Flux<FCM> findByUserId(long id);
 }
